@@ -15,57 +15,62 @@ import (
 	"path/filepath"
 	"sort"
 	"sync"
-	"sync/atomic"
 	"time"
 )
 
 var (
-	StopError = errors.New("state")
-	NotLeader = errors.New("not leader")
+	StopError        = errors.New("state")
+	NotLeaderError   = errors.New("not leader")
+	NotFollowerError = errors.New("not follower")
+	NotSnapshotError = errors.New("not snapshot")
 )
 
+type event struct {
+	target   any
+	retValue any
+	c        chan error
+}
+
 type server struct {
-	grpcServer            *grpc.Server                // grpc server
-	currentTerm           uint64                      // 当前任期
-	locker                sync.RWMutex                // 读写锁
-	config                *Config                     // 配置文件
-	stateMachine          StateMachine                // 状态机
-	pendingSnapshot       *Snapshot                   // 待存储的快照
-	snapshot              *Snapshot                   // 最新快照
-	votedFor              uint64                      // 选票投给了哪个候选人
-	log                   *Log                        // 日志对象
-	memberMgr             *memberManager              // 集群中节点的管理（不包括自身节点）
-	state                 pb.NodeState                // 节点的当前状态,follower/candidate/leaderId/snapshotting/state/initialized
-	stopped               chan struct{}               // 停止信号
-	routineGroup          sync.WaitGroup              // 保证协程能够安全退出
-	leaderId              uint64                      // 当state为follower时，设置为leader的id
-	followerHeartbeatChan chan bool                   // follower用于更新与leader之间的心跳
-	leaderRespChan        chan *AppendEntriesResponse // leaderId loop use
-	cmdChan               chan *Command
-	cmdPool               *sync.Pool
+	grpcServer      *grpc.Server   // grpc server
+	currentTerm     uint64         // 当前任期
+	locker          sync.RWMutex   // 读写锁
+	config          *Config        // 配置文件
+	stateMachine    StateMachine   // 状态机
+	pendingSnapshot *Snapshot      // 待存储的快照
+	snapshot        *Snapshot      // 最新快照
+	votedFor        uint64         // 选票投给了哪个候选人
+	log             *Log           // 日志对象
+	members         *memberManager // 集群中节点的管理（不包括自身节点）
+	state           pb.NodeState   // 节点的当前状态,follower/candidate/leaderId/snapshotting/state/initialized
+	stopped         chan struct{}  // 停止信号
+	routineGroup    sync.WaitGroup // 保证协程能够安全退出
+	leaderId        uint64         // 当state为follower时，设置为leader的id
+	cmdChan         chan *Command
+	cmdPool         *sync.Pool
+	c               chan *event
 }
 
 func New(config *Config, stateMachine StateMachine) Raft {
-	pool := sync.Pool{New: func() interface{} {
+	pool := sync.Pool{New: func() any {
 		return new(Command)
 	}}
 	return &server{
-		grpcServer:            nil,
-		locker:                sync.RWMutex{},
-		config:                config,
-		stateMachine:          stateMachine,
-		pendingSnapshot:       nil,
-		snapshot:              nil,
-		votedFor:              0,
-		log:                   newLog(config.WriteType, config.LogWriteTimeInterval, stateMachine.Apply),
-		memberMgr:             newMemberManager(),
-		stopped:               make(chan struct{}, 1),
-		routineGroup:          sync.WaitGroup{},
-		leaderId:              0,
-		followerHeartbeatChan: make(chan bool, 256),
-		leaderRespChan:        make(chan *AppendEntriesResponse, 1024),
-		cmdChan:               make(chan *Command, 1024),
-		cmdPool:               &pool,
+		grpcServer:      nil,
+		locker:          sync.RWMutex{},
+		config:          config,
+		stateMachine:    stateMachine,
+		pendingSnapshot: nil,
+		snapshot:        nil,
+		votedFor:        0,
+		log:             newLog(config.WriteType, config.LogWriteTimeInterval, stateMachine.Apply),
+		members:         newMemberManager(),
+		stopped:         make(chan struct{}, 1),
+		routineGroup:    sync.WaitGroup{},
+		leaderId:        0,
+		cmdChan:         make(chan *Command, 1024),
+		cmdPool:         &pool,
+		c:               make(chan *event, 4096),
 	}
 }
 
@@ -133,51 +138,19 @@ func (s *server) Init() error {
 	return nil
 }
 
-func (s *server) RequestVote(ctx context.Context, req *pb.RequestVoteRequest) (*pb.RequestVoteResponse, error) {
-	log.Debug("RequestVote")
-
-	if req.Term < s.CurrentTerm() {
-		return &pb.RequestVoteResponse{Term: s.CurrentTerm(), VoteGranted: false}, nil
-	}
-
-	if req.Term > s.CurrentTerm() {
-		// 任期比对方小，成为跟随者
-		log.Debugf("req.CurrentTerm > s.CurrentTerm()")
-		s.updateCurrentTerm(req.Term, 0)
-	} else if atomic.LoadUint64(&s.votedFor) != 0 && atomic.LoadUint64(&s.votedFor) != req.CandidateId {
-		// 已投过票，且投票对象不是req.CandidateId
-		log.Debugf("have been voted, votedFor:%d candidateId:%d", s.votedFor, req.CandidateId)
-		return &pb.RequestVoteResponse{Term: s.CurrentTerm(), VoteGranted: false}, nil
-	}
-
-	lastLogIndex, lastLogTerm := s.log.lastInfo()
-	if lastLogIndex > req.LastLogIndex || lastLogTerm > req.LastLogTerm {
-		// 当前节点的日志更新，不做投票
-		log.Debugf("local log is more newer")
-		return &pb.RequestVoteResponse{Term: s.CurrentTerm(), VoteGranted: false}, nil
-	}
-
-	// 选票投给 req.CandidateId
-	atomic.SwapUint64(&s.votedFor, req.CandidateId)
-	s.followerHeartbeatChan <- true
-	log.Infof("votedFor: %d", s.votedFor)
-
-	return &pb.RequestVoteResponse{Term: s.CurrentTerm(), VoteGranted: true}, nil
-}
-
 func (s *server) Join(existing string) {
 	if existing == "" || len(existing) == 0 {
 		return
 	}
 
-	join := func(addr string) *pb.MemberResponse {
+	join := func(addr string) *pb.AddMemberResponse {
 		cc, err := dial(addr, s.config.DialTimeout, s.config.DialOptions)
 		if err != nil {
 			log.Errorf("dial %s failed: %v", addr, err)
 			return nil
 		}
 
-		req := &pb.MemberRequest{
+		req := &pb.AddMemberRequest{
 			Leader: false,
 			Member: pb.Member{Id: s.ID(), Address: s.Address()},
 		}
@@ -216,7 +189,7 @@ func (s *server) Do(commandName string, command []byte) <-chan error {
 
 	go func() {
 		if !s.IsLeader() {
-			done <- NotLeader
+			done <- NotLeaderError
 			return
 		}
 
@@ -264,51 +237,174 @@ func (s *server) Do(commandName string, command []byte) <-chan error {
 
 // AppendEntries 用于leader进行日志复制与心跳检测
 func (s *server) AppendEntries(ctx context.Context, req *pb.AppendEntriesRequest) (*pb.AppendEntriesResponse, error) {
-	// log.Debugf("AppendEntries, request: %v", *req)
+	log.Debugf("AppendEntries, request: %v", *req)
 	//since := time.Now()
 	//defer func() {
 	//	log.Infof("time used:%d ms", time.Now().Sub(since).Milliseconds())
 	//}()
+	resp, err := s.send(req)
+	if err != nil {
+		return &pb.AppendEntriesResponse{}, err
+	}
+	return resp.(*pb.AppendEntriesResponse), nil
+}
 
+func (s *server) RequestVote(ctx context.Context, req *pb.RequestVoteRequest) (*pb.RequestVoteResponse, error) {
+	log.Debug("RequestVote")
+	resp, err := s.send(req)
+	if err != nil {
+		return &pb.RequestVoteResponse{}, err
+	}
+	return resp.(*pb.RequestVoteResponse), nil
+}
+
+func (s *server) SnapshotAsk(ctx context.Context, req *pb.SnapshotAskRequest) (*pb.SnapshotAskResponse, error) {
+	log.Debug("SnapshotAsk")
+	switch s.State() {
+	case pb.NodeState_Follower:
+		resp, err := s.send(req)
+		if err != nil {
+			return &pb.SnapshotAskResponse{}, err
+		}
+		return resp.(*pb.SnapshotAskResponse), nil
+	default:
+		return &pb.SnapshotAskResponse{Success: false}, NotFollowerError
+	}
+}
+
+func (s *server) SnapshotRecovery(ctx context.Context, req *pb.SnapshotRecoveryRequest) (*pb.SnapshotRecoveryResponse, error) {
+	log.Debug("SnapshotRecovery")
+	switch s.State() {
+	case pb.NodeState_Snapshotting:
+		resp, err := s.send(req)
+		if err != nil {
+			return &pb.SnapshotRecoveryResponse{}, err
+		}
+		return resp.(*pb.SnapshotRecoveryResponse), nil
+	default:
+		return &pb.SnapshotRecoveryResponse{Success: false}, NotSnapshotError
+	}
+}
+
+// Membership 成员信息,由leader->follower
+func (s *server) Membership(ctx context.Context, req *pb.MembershipRequest) (*pb.MembershipResponse, error) {
+	log.Debug("Membership")
+	for _, m := range req.Members {
+		if m.Id == s.ID() {
+			continue
+		}
+		if s.members.Exist(m.Id) {
+			continue
+		}
+		s.members.Store(newMember(m, s))
+		log.Infof("add member %d %s", m.Id, m.Address)
+	}
+	return &pb.MembershipResponse{Success: true}, nil
+}
+
+// AddMember 添加成员，只能由leader完成
+func (s *server) AddMember(ctx context.Context, req *pb.AddMemberRequest) (*pb.AddMemberResponse, error) {
+	log.Debug("AddMember")
+	resp, err := s.send(req)
+	if err != nil {
+		return &pb.AddMemberResponse{}, err
+	}
+	return resp.(*pb.AddMemberResponse), nil
+}
+
+// RemoveMember 移除节点，只能由leader完成
+func (s *server) RemoveMember(ctx context.Context, req *pb.RemoveMemberRequest) (*pb.RemoveMemberResponse, error) {
+	log.Debug("RemoveMember")
+	resp, err := s.send(req)
+	if err != nil {
+		return &pb.RemoveMemberResponse{}, err
+	}
+	return resp.(*pb.RemoveMemberResponse), nil
+}
+
+// Ping ping message
+func (s *server) Ping(ctx context.Context, req *pb.Empty) (*pb.Empty, error) {
+	log.Debug("Ping")
+	return &pb.Empty{}, nil
+}
+
+func (s *server) send(v any) (any, error) {
 	if !s.Running() {
 		return nil, StopError
 	}
 
+	e := &event{target: v, c: make(chan error, 1)}
+
+	select {
+	case <-s.stopped:
+		return nil, StopError
+	case s.c <- e:
+	}
+
+	select {
+	case <-s.stopped:
+		return nil, StopError
+	case err := <-e.c:
+		return e.retValue, err
+	}
+}
+
+func (s *server) sendAsync(v any) {
+	if !s.Running() {
+		return
+	}
+
+	e := &event{target: v, c: make(chan error, 1)}
+
+	select {
+	case s.c <- e:
+		return
+	default:
+	}
+
+	s.routineGroup.Add(1)
+	go func() {
+		defer s.routineGroup.Done()
+		select {
+		case s.c <- e:
+		case <-s.stopped:
+		}
+	}()
+}
+
+func (s *server) processAppendEntriesRequest(req *pb.AppendEntriesRequest) (*pb.AppendEntriesResponse, bool) {
 	// 1、reply false if term < currentTerm
-	currentTerm := s.CurrentTerm()
-	if req.Term < currentTerm {
+	if req.Term < s.currentTerm {
 		log.Debug("stale term")
 		return &pb.AppendEntriesResponse{
-			Term:        s.CurrentTerm(),
+			Term:        s.currentTerm,
 			Index:       s.log.CurrentIndex(),
 			CommitIndex: s.log.CommitIndex(),
 			Success:     false,
-		}, nil
-	} else if req.Term == currentTerm {
-		if s.State() == pb.NodeState_Candidate {
+		}, false
+
+	} else if req.Term == s.currentTerm {
+		if s.state == pb.NodeState_Candidate {
 			s.setState(pb.NodeState_Follower)
 		}
-		s.locker.Lock()
 		s.leaderId = req.LeaderId
-		s.locker.Unlock()
+
 	} else {
 		// update term and leaderId
 		log.Debug("update currentTerm, become follower")
 		s.updateCurrentTerm(req.Term, req.LeaderId)
 	}
 
-	s.followerHeartbeatChan <- true
-
 	// 2、reply false if log doesn't contain an Entry at prevLogIndex whose
 	// term matches prevLogTerm
 	if err := s.log.truncate(req.PrevLogIndex, req.PrevLogTerm); err != nil {
 		log.Debug(err)
 		return &pb.AppendEntriesResponse{
-			Term:        s.CurrentTerm(),
+			Term:        s.currentTerm,
 			Index:       s.log.CurrentIndex(),
 			CommitIndex: s.log.CommitIndex(),
 			Success:     false,
-		}, nil
+		}, true
 	}
 
 	// 3、if an existing Entry conflicts with a new one(same index but
@@ -317,11 +413,11 @@ func (s *server) AppendEntries(ctx context.Context, req *pb.AppendEntriesRequest
 	if err := s.log.appendEntries(req.Entries); err != nil {
 		log.Warn(err)
 		return &pb.AppendEntriesResponse{
-			Term:        s.CurrentTerm(),
+			Term:        s.currentTerm,
 			Index:       s.log.CurrentIndex(),
 			CommitIndex: s.log.CommitIndex(),
 			Success:     false,
-		}, nil
+		}, true
 	}
 
 	log.Debugf("append entries: (prevLogIndex:%d entries:%d)",
@@ -330,52 +426,161 @@ func (s *server) AppendEntries(ctx context.Context, req *pb.AppendEntriesRequest
 	if err := s.log.setCommitIndex(req.CommitIndex); err != nil {
 		log.Warn(err)
 		return &pb.AppendEntriesResponse{
-			Term:        s.CurrentTerm(),
+			Term:        s.currentTerm,
 			Index:       s.log.CurrentIndex(),
 			CommitIndex: s.log.CommitIndex(),
 			Success:     false,
-		}, nil
+		}, true
 	}
 
 	// log.Debugf("AppendEntries success")
 	// 若返回成功，则代表leader和follower的日志条目一定相同
 	return &pb.AppendEntriesResponse{
-		Term:        s.CurrentTerm(),
+		Term:        s.currentTerm,
 		Index:       s.log.CurrentIndex(),
 		CommitIndex: s.log.CommitIndex(),
 		Success:     true,
-	}, nil
+	}, true
 }
 
-func (s *server) SnapshotAsk(ctx context.Context, req *pb.SnapshotAskRequest) (*pb.SnapshotAskResponse, error) {
-	log.Debug("SnapshotAsk")
+func (s *server) processAppendEntriesResponse(resp *AppendEntriesResponse, syncedMember map[uint64]struct{}) {
+	if resp.Term > s.currentTerm {
+		// 主动退位
+		log.Debug("resp.CurrentTerm > s.CurrentTerm(), change from leader to follower")
+		s.updateCurrentTerm(resp.Term, 0)
+		return
+	}
 
-	resp := &pb.SnapshotAskResponse{Success: false}
+	if !resp.Success {
+		return
+	}
 
+	if resp.Index <= s.log.CommitIndex() {
+		return
+	}
+
+	switch s.config.ReplicationType {
+	case Semisynchronous:
+		_ = s.log.setCommitIndex(resp.Index)
+
+	case Synchronous:
+		syncedMember[resp.Id] = struct{}{}
+
+		// 检查超过半数的条件
+		if len(syncedMember)+1 < s.QuorumSize() {
+			return
+		}
+
+		// 找到(1/2+1)个节点都确认的日志条目索引
+		var indices []uint64
+		indices = append(indices, s.log.CurrentIndex())
+		s.members.Range(func(m *member) bool {
+			indices = append(indices, m.getPrevLogIndex())
+			return true
+		})
+
+		sort.Sort(sort.Reverse(Uint64Slice(indices))) // 从大到小排序
+		commitIndex := indices[s.QuorumSize()-1]
+
+		// 更新 commitIndex
+		_ = s.log.setCommitIndex(commitIndex)
+
+		// 重置
+		syncedMember = make(map[uint64]struct{}, s.MemberCount())
+	}
+}
+
+func (s *server) processRequestVoteRequest(req *pb.RequestVoteRequest) (*pb.RequestVoteResponse, bool) {
+	if req.Term < s.currentTerm {
+		return &pb.RequestVoteResponse{Term: s.currentTerm, VoteGranted: false}, false
+	}
+
+	if req.Term > s.currentTerm {
+		// 任期比对方小，成为跟随者
+		log.Debugf("req.CurrentTerm > s.CurrentTerm()")
+		s.updateCurrentTerm(req.Term, 0)
+
+	} else if s.votedFor != 0 && s.votedFor != req.CandidateId {
+		// 已投过票，且投票对象不是req.CandidateId
+		log.Debugf("have been voted, votedFor:%d candidateId:%d", s.votedFor, req.CandidateId)
+		return &pb.RequestVoteResponse{Term: s.currentTerm, VoteGranted: false}, false
+	}
+
+	lastLogIndex, lastLogTerm := s.log.lastInfo()
+	if lastLogIndex > req.LastLogIndex || lastLogTerm > req.LastLogTerm {
+		// 当前节点的日志更新，不做投票
+		log.Debugf("local log is more newer")
+		return &pb.RequestVoteResponse{Term: s.currentTerm, VoteGranted: false}, false
+	}
+
+	// 选票投给 req.CandidateId
+	s.votedFor = req.CandidateId
+	log.Infof("votedFor: %d", s.votedFor)
+
+	return &pb.RequestVoteResponse{Term: s.currentTerm, VoteGranted: true}, true
+}
+
+func (s *server) processRequestVoteResponse(resp *pb.RequestVoteResponse) bool {
+	if resp.VoteGranted && resp.Term == s.currentTerm {
+		return true
+	}
+
+	if resp.Term > s.currentTerm {
+		log.Debug("resp.CurrentTerm > s.CurrentTerm")
+		s.updateCurrentTerm(resp.Term, 0)
+	} else {
+		// 节点拒绝了对当前节点的投票
+	}
+
+	return false
+}
+
+func (s *server) processAddMemberRequest(req *pb.AddMemberRequest) (*pb.AddMemberResponse, bool) {
+	if !s.IsLeader() && !req.Leader {
+		return &pb.AddMemberResponse{
+			Success: false,
+			Leader:  pb.Member{Id: s.LeaderId(), Address: s.LeaderAddress()},
+		}, false
+	}
+
+	s.addMember(req.Member)
+	return &pb.AddMemberResponse{Success: true}, true
+}
+
+func (s *server) processRemoveMemberRequest(req *pb.RemoveMemberRequest) (*pb.RemoveMemberResponse, bool) {
+	if s.state != pb.NodeState_Leader && !req.Leader {
+		return &pb.RemoveMemberResponse{
+			Success: false,
+			Leader:  pb.Member{Id: s.LeaderId(), Address: s.LeaderAddress()},
+		}, false
+	}
+
+	s.removeMember(req.Member.Id)
+	return &pb.RemoveMemberResponse{Success: true}, true
+}
+
+func (s *server) processSnapshotAskRequest(req *pb.SnapshotAskRequest) *pb.SnapshotAskResponse {
 	entry := s.log.getEntry(req.LastIndex)
 	if entry != nil && entry.Term == req.LastTerm {
-		return resp, nil
+		return &pb.SnapshotAskResponse{Success: false}
 	}
 
 	s.setState(pb.NodeState_Snapshotting)
-	resp.Success = true
-	return resp, nil
+	return &pb.SnapshotAskResponse{Success: true}
 }
 
-func (s *server) SnapshotRecovery(ctx context.Context, req *pb.SnapshotRecoveryRequest) (*pb.SnapshotRecoveryResponse, error) {
-	log.Debug("SnapshotRecovery")
-
+func (s *server) processSnapshotRecoveryRequest(req *pb.SnapshotRecoveryRequest) *pb.SnapshotRecoveryResponse {
 	if err := s.stateMachine.Recovery(req.State); err != nil {
 		log.Fatal("cannot recover from previous state")
 	}
-
-	s.currentTerm = req.LastTerm
-	s.log.updateCommitIndex(req.LastIndex)
 
 	// add member
 	for _, mb := range req.Members {
 		s.addMember(mb)
 	}
+
+	s.currentTerm = req.LastTerm
+	s.log.updateCommitIndex(req.LastIndex)
 
 	// 创建本地快照
 	s.pendingSnapshot = newSnapshot(s.snapshotPath(req.LastIndex, req.LastTerm), &pb.Snapshot{
@@ -392,58 +597,7 @@ func (s *server) SnapshotRecovery(ctx context.Context, req *pb.SnapshotRecoveryR
 	log.Debugf("update currentTerm, become follower")
 	s.updateCurrentTerm(req.LastTerm, req.LeaderId)
 
-	return &pb.SnapshotRecoveryResponse{
-		Term:        s.CurrentTerm(),
-		Success:     true,
-		CommitIndex: s.log.CommitIndex(),
-	}, nil
-}
-
-// Membership 成员信息,由leader->follower
-func (s *server) Membership(ctx context.Context, req *pb.MembershipRequest) (*pb.MembershipResponse, error) {
-	for _, m := range req.Members {
-		if m.Id == s.ID() {
-			continue
-		}
-		if s.memberMgr.Exist(m.Id) {
-			continue
-		}
-		s.memberMgr.Store(newMember(m, s))
-		log.Infof("add member %d %s", m.Id, m.Address)
-	}
-	return &pb.MembershipResponse{Success: true}, nil
-}
-
-// AddMember 添加成员，只能由leader完成
-func (s *server) AddMember(ctx context.Context, req *pb.MemberRequest) (*pb.MemberResponse, error) {
-	if !s.IsLeader() && !req.Leader {
-		return &pb.MemberResponse{
-			Success: false,
-			Leader:  pb.Member{Id: s.LeaderId(), Address: s.LeaderAddress()},
-		}, nil
-	}
-
-	s.addMember(req.Member)
-	return &pb.MemberResponse{Success: true}, nil
-}
-
-// RemoveMember 移除节点，只能由leader完成
-func (s *server) RemoveMember(ctx context.Context, req *pb.MemberRequest) (*pb.MemberResponse, error) {
-	if !s.IsLeader() && !req.Leader {
-		return &pb.MemberResponse{
-			Success: false,
-			Leader:  pb.Member{Id: s.LeaderId(), Address: s.LeaderAddress()},
-		}, nil
-	}
-
-	s.removeMember(req.Member.Id)
-	return &pb.MemberResponse{Success: true}, nil
-}
-
-// Ping ping message
-func (s *server) Ping(ctx context.Context, req *pb.Empty) (*pb.Empty, error) {
-	// log.Debug("Ping")
-	return &pb.Empty{}, nil
+	return &pb.SnapshotRecoveryResponse{Term: req.LastTerm, Success: true, CommitIndex: req.LastIndex}
 }
 
 func (s *server) Stop() {
@@ -464,14 +618,17 @@ func (s *server) Stop() {
 	s.log.close()
 	s.setState(pb.NodeState_Stopped)
 
+	close(s.cmdChan)
+	close(s.c)
+
 	log.Info("stopped")
 }
 
 // leaveCluster 离开集群
 func (s *server) leaveCluster() {
 	if s.IsLeader() {
-		s.memberMgr.Range(func(m *member) bool {
-			_ = m.sendRemoveMemberRequest(&pb.MemberRequest{Leader: true, Member: s.Self()})
+		s.members.Range(func(m *member) bool {
+			_ = m.sendRemoveMemberRequest(&pb.RemoveMemberRequest{Leader: true, Member: s.Self()})
 			return true
 		})
 		return
@@ -481,12 +638,12 @@ func (s *server) leaveCluster() {
 		return
 	}
 
-	leader, ok := s.memberMgr.Load(s.leaderId)
+	leader, ok := s.members.Load(s.leaderId)
 	if !ok {
 		return
 	}
 
-	_ = leader.sendRemoveMemberRequest(&pb.MemberRequest{Leader: false, Member: s.Self()})
+	_ = leader.sendRemoveMemberRequest(&pb.RemoveMemberRequest{Leader: false, Member: s.Self()})
 }
 
 func (s *server) IsLeader() bool {
@@ -511,7 +668,7 @@ func (s *server) LeaderAddress() string {
 	case pb.NodeState_Leader:
 		return s.Address()
 	case pb.NodeState_Follower:
-		m, _ := s.memberMgr.Load(s.leaderId)
+		m, _ := s.members.Load(s.leaderId)
 		return m.Address
 	default:
 		return ""
@@ -554,23 +711,22 @@ func (s *server) updateCurrentTerm(term uint64, leaderId uint64) {
 
 	if s.State() == pb.NodeState_Leader {
 		// 如果是leader，则停止所有心跳
-		s.memberMgr.Range(func(m *member) bool {
+		s.members.Range(func(m *member) bool {
 			m.stopHeartbeat(false)
 			return true
 		})
 	}
 
-	if s.State() != pb.NodeState_Follower {
+	if s.state != pb.NodeState_Follower {
 		s.setState(pb.NodeState_Follower)
 	}
 
 	// 更新状态
 	s.locker.Lock()
-	defer s.locker.Unlock()
-
 	s.currentTerm = term
 	s.leaderId = leaderId
 	s.votedFor = 0
+	s.locker.Unlock()
 }
 
 // loop 事件循环
@@ -603,8 +759,23 @@ func (s *server) followerLoop() {
 		case <-s.stopped:
 			s.setState(pb.NodeState_Stopped)
 			return
-		case update = <-s.followerHeartbeatChan:
-			// 收到leader的心跳消息，准备更新超时时间
+
+		case e := <-s.c:
+			var err error
+			switch msg := e.target.(type) {
+			case *pb.AppendEntriesRequest:
+				e.retValue, update = s.processAppendEntriesRequest(msg)
+			case *pb.RequestVoteRequest:
+				e.retValue, update = s.processRequestVoteRequest(msg)
+			case *pb.AddMemberRequest:
+				e.retValue, update = s.processAddMemberRequest(msg)
+			case *pb.RemoveMemberRequest:
+				e.retValue, update = s.processRemoveMemberRequest(msg)
+			case *pb.SnapshotAskRequest:
+				e.retValue = s.processSnapshotAskRequest(msg)
+			}
+			e.c <- err
+
 		case <-timeoutChan:
 			// 超时未收到leader的心跳消息
 			if s.log.CurrentIndex() > 0 || s.MemberCount() == 1 {
@@ -649,7 +820,7 @@ func (s *server) candidateLoop() {
 			}
 
 			// 向连接的节点发送请求选票的消息
-			s.memberMgr.Range(func(m *member) bool {
+			s.members.Range(func(m *member) bool {
 				s.routineGroup.Add(1)
 				go func(m *member) {
 					defer s.routineGroup.Done()
@@ -676,18 +847,26 @@ func (s *server) candidateLoop() {
 		case <-s.stopped:
 			s.setState(pb.NodeState_Stopped)
 			return
+
 		case resp := <-respChan:
-			if resp.VoteGranted && resp.Term == s.CurrentTerm() {
+			if succ := s.processRequestVoteResponse(resp); succ {
 				votesGranted++
-				log.Debugf("votesGranted=%d quorumSize=%d", votesGranted, s.QuorumSize())
-				break
 			}
-			if resp.Term > s.CurrentTerm() {
-				log.Debug("resp.CurrentTerm > s.CurrentTerm")
-				s.updateCurrentTerm(resp.Term, 0)
-			} else {
-				// 节点拒绝了对当前节点的投票
+
+		case e := <-s.c:
+			var err error
+			switch msg := e.target.(type) {
+			case *pb.AppendEntriesRequest:
+				e.retValue, _ = s.processAppendEntriesRequest(msg)
+			case *pb.RequestVoteRequest:
+				e.retValue, _ = s.processRequestVoteRequest(msg)
+			case *pb.AddMemberRequest:
+				e.retValue, _ = s.processAddMemberRequest(msg)
+			case *pb.RemoveMemberRequest:
+				e.retValue, _ = s.processRemoveMemberRequest(msg)
 			}
+			e.c <- err
+
 		case <-timeoutChan:
 			// 选举超时，开始新一轮选举
 			log.Infof("candidate vote timeout")
@@ -699,9 +878,7 @@ func (s *server) candidateLoop() {
 // leaderLoop
 // 1、成员保活 2、发送日志条目 3、日志提交
 func (s *server) leaderLoop() {
-	memberCount := s.MemberCount()
-	s.leaderRespChan = make(chan *AppendEntriesResponse, memberCount)
-	s.memberMgr.Range(func(m *member) bool {
+	s.members.Range(func(m *member) bool {
 		m.setPrevLogIndex(s.log.CurrentIndex())
 		m.startHeartbeat()
 		return true
@@ -723,12 +900,12 @@ func (s *server) leaderLoop() {
 	}()
 
 	// 已同步了commit后的日志条目的节点数（）
-	syncedMember := make(map[uint64]struct{}, s.MemberCount())
+	syncedMember := make(map[uint64]struct{})
 
 	for s.State() == pb.NodeState_Leader {
 		select {
 		case <-s.stopped:
-			s.memberMgr.Range(func(m *member) bool {
+			s.members.Range(func(m *member) bool {
 				m.stopHeartbeat(false)
 				return true
 			})
@@ -740,7 +917,7 @@ func (s *server) leaderLoop() {
 
 		case cmd := <-s.cmdChan:
 			entry := &pb.LogEntry{
-				Term:        s.CurrentTerm(),
+				Term:        s.currentTerm,
 				Index:       s.log.nextIndex(),
 				CommandName: cmd.name,
 				Command:     cmd.data,
@@ -763,58 +940,21 @@ func (s *server) leaderLoop() {
 				_ = s.log.setCommitIndex(entry.Index)
 			}
 
-		case resp := <-s.leaderRespChan:
-			if resp.Term > s.currentTerm {
-				// 主动退位
-				log.Debug("resp.CurrentTerm > s.CurrentTerm(), change from leader to follower")
-				s.updateCurrentTerm(resp.Term, 0)
-				break
+		case e := <-s.c:
+			var err error
+			switch msg := e.target.(type) {
+			case *pb.AppendEntriesRequest:
+				e.retValue, _ = s.processAppendEntriesRequest(msg)
+			case *AppendEntriesResponse:
+				s.processAppendEntriesResponse(msg, syncedMember)
+			case *pb.RequestVoteRequest:
+				e.retValue, _ = s.processRequestVoteRequest(msg)
+			case *pb.AddMemberRequest:
+				e.retValue, _ = s.processAddMemberRequest(msg)
+			case *pb.RemoveMemberRequest:
+				e.retValue, _ = s.processRemoveMemberRequest(msg)
 			}
-			if !resp.Success {
-				break
-			}
-			if resp.Index <= s.log.CommitIndex() {
-				break
-			}
-
-			switch s.config.ReplicationType {
-			case Synchronous:
-				syncedMember[resp.Id] = struct{}{}
-
-				// 检查超过半数的条件
-				if len(syncedMember)+1 < s.QuorumSize() {
-					continue
-				}
-
-				// 找到(1/2+1)个节点都确认的日志条目索引
-				var indices []uint64
-				indices = append(indices, s.log.CurrentIndex())
-				s.memberMgr.Range(func(m *member) bool {
-					indices = append(indices, m.getPrevLogIndex())
-					return true
-				})
-
-				sort.Sort(sort.Reverse(Uint64Slice(indices))) // 从大到小排序
-				commitIndex := indices[s.QuorumSize()-1]
-
-				// 更新 commitIndex
-				s.locker.Lock()
-				_ = s.log.setCommitIndex(commitIndex)
-				s.locker.Unlock()
-
-				// 重置
-				syncedMember = make(map[uint64]struct{}, s.MemberCount())
-
-			case Semisynchronous:
-				s.locker.Lock()
-				_ = s.log.setCommitIndex(resp.Index)
-				s.locker.Unlock()
-
-			case Asynchronous:
-
-			default:
-				log.Error("unknown replication type")
-			}
+			e.c <- err
 		}
 	}
 }
@@ -825,24 +965,29 @@ func (s *server) snapshotLoop() {
 		case <-s.stopped:
 			s.setState(pb.NodeState_Stopped)
 			return
-		default:
+
+		case e := <-s.c:
+			var err error
+			switch msg := e.target.(type) {
+			case *pb.AppendEntriesRequest:
+				e.retValue, _ = s.processAppendEntriesRequest(msg)
+			case *pb.RequestVoteRequest:
+				e.retValue, _ = s.processRequestVoteRequest(msg)
+			case *pb.AddMemberRequest:
+				e.retValue, _ = s.processAddMemberRequest(msg)
+			case *pb.RemoveMemberRequest:
+				e.retValue, _ = s.processRemoveMemberRequest(msg)
+			case *pb.SnapshotRecoveryRequest:
+				e.retValue = s.processSnapshotRecoveryRequest(msg)
+			}
+			e.c <- err
 		}
 	}
 }
 
-//// voteFor 投票，如果已投过票，则投票失败
-//func (s *server) voteFor(id uint64) bool {
-//	return atomic.CompareAndSwapUint64(&s.votedFor, 0, id)
-//}
-//
-//// resetVote 重置选票（未投票状态）
-//func (s *server) resetVote() {
-//	atomic.SwapUint64(&s.votedFor, 0)
-//}
-
 func (s *server) addMember(mem pb.Member) {
 	// 不允许节点加入两次
-	if s.memberMgr.Exist(mem.Id) {
+	if s.members.Exist(mem.Id) {
 		return
 	}
 
@@ -853,7 +998,7 @@ func (s *server) addMember(mem pb.Member) {
 	m := newMember(mem, s)
 
 	// 再次检查是否添加成功
-	if ok := s.memberMgr.Store(m); !ok {
+	if ok := s.members.Store(m); !ok {
 		return
 	}
 	log.Infof("add member %d %s", m.Id, m.Address)
@@ -873,11 +1018,11 @@ func (s *server) addMember(mem pb.Member) {
 		}()
 
 		// 将新节点加入的信息通知给其它节点
-		req := &pb.MemberRequest{
+		req := &pb.AddMemberRequest{
 			Leader: true,
 			Member: mem,
 		}
-		s.memberMgr.Range(func(m *member) bool {
+		s.members.Range(func(m *member) bool {
 			go func(m *member) {
 				_ = m.sendAddMemberRequest(req)
 			}(m)
@@ -891,22 +1036,22 @@ func (s *server) removeMember(id uint64) {
 		return
 	}
 
-	m, loaded := s.memberMgr.Delete(id)
+	m, loaded := s.members.Delete(id)
 	if !loaded {
 		return
 	}
 
 	log.Infof("remove member %d", id)
 
-	if s.IsLeader() {
+	if s.state == pb.NodeState_Leader {
 		s.routineGroup.Add(1)
 		go func() {
 			defer s.routineGroup.Done()
 			m.stopHeartbeat(true)
 		}()
 
-		req := &pb.MemberRequest{Leader: true, Member: m.Member}
-		s.memberMgr.Range(func(m *member) bool {
+		req := &pb.RemoveMemberRequest{Leader: true, Member: m.Member}
+		s.members.Range(func(m *member) bool {
 			go func(m *member) {
 				_ = m.sendRemoveMemberRequest(req)
 			}(m)
@@ -917,7 +1062,7 @@ func (s *server) removeMember(id uint64) {
 
 func (s *server) Members() []pb.Member {
 	members := []pb.Member{s.Self()}
-	s.memberMgr.Range(func(m *member) bool {
+	s.members.Range(func(m *member) bool {
 		members = append(members, m.Member)
 		return true
 	})
@@ -925,7 +1070,7 @@ func (s *server) Members() []pb.Member {
 }
 
 func (s *server) MemberCount() int {
-	return s.memberMgr.Count() + 1
+	return s.members.Count() + 1
 }
 
 // QuorumSize 日志提交需要超过半数的节点（n/2+1）
@@ -1023,7 +1168,7 @@ func (s *server) LoadSnapshot() error {
 	filenames, err := dir.Readdirnames(-1)
 	if err != nil {
 		_ = dir.Close()
-		panic(err)
+		panic(any(err))
 	}
 	_ = dir.Close()
 
